==> Subreddits NLP/01_Getting_subreddits.ipynb <==
%%javascript
IPython.OutputArea.prototype._should_scroll = function(lines) {
return false;
}

pd.set_option('display.min_rows', 60)

import requests
import time
import datetime as dt

### We used this function (code by Josh Robin) to extract the subreddits. We changed some parameters to extract more posts in one sitting. 

def query_pushshift(subreddit, kind='submission', skip=20, times=150, 
                    subfield = ['title', 'selftext', 'subreddit', 'created_utc', 'author', 'num_comments', 'score', 'is_self'],
                    comfields = ['body', 'score', 'created_utc']):

    stem = "https://api.pushshift.io/reddit/search/{}/?subreddit={}&size=1000".format(kind, subreddit)
    mylist = []
    
    for x in range(1, times + 1):
        
        URL = "{}&after={}d".format(stem, skip * x)
        print(URL)
        response = requests.get(URL)
        assert response.status_code == 200
        mine = response.json()['data']
        df = pd.DataFrame.from_dict(mine)
        mylist.append(df)
        time.sleep(2)
        
    full = pd.concat(mylist, sort=False)
    
    if kind == "submission":
        
        full = full[subfield]
        
        full = full.drop_duplicates()
        
        full = full.loc[full['is_self'] == True]
        
    def get_date(created):
        return dt.date.fromtimestamp(created)
    
    _timestamp = full["created_utc"].apply(get_date)
    
    full['timestamp'] = _timestamp

    print(full.shape)
    
    return full 

math_total = query_pushshift('math')

physics_total = query_pushshift('physics')

ask_physics_total = query_pushshift('askphysics')

#### Done with getting the data with no duplicates

# physics_total.duplicated().value_counts()   

# ask_physics_total.duplicated().value_counts()

#### Saving the data to csv files for later use:

# math_total.to_csv('math.csv', index=False)
# physics_total.to_csv('physics.csv', index=False)
# ask_physics_total.to_csv('askphysics.csv', index=False)
