==> Subreddits NLP/02_Joining_and_EDA.ipynb <==
### Reading the subreddits, joining them with concat and doing some EDA

math = pd.read_csv('./data/math.csv')

physics = pd.read_csv('./data/physics.csv')

askphysics = pd.read_csv('./data/askphysics.csv')

frames = [math, physics, askphysics]   # Creating a list of dataframes

df = pd.concat(frames, ignore_index=True)   # Concatinating dataframes

# Dropping unnecesary columns 
data = df.drop(columns=['created_utc', 'author', 'num_comments', 'score', 'is_self', 'timestamp'])

data['post'] = data['title'] +' '+ data['selftext']  # joining title and text

data['post'][0]

data = data.drop(columns=['title', 'selftext'])

data

data.dropna(inplace=True)   # Getting rid of nulls

data

data['subreddit'].value_counts()   

# Changing target column to binary
data['subreddit'] = data['subreddit'].map({'math':1, 'Physics':0, 'AskPhysics':0})

data['subreddit'].value_counts()

data.isnull().sum()   # Checking subreddits that got through the cracks

data.dropna(inplace=True)    # Dropping subreddits that got through the cracks

data

data.isnull().sum()

data   # Final version of data

data.to_csv('data/data.csv', index=False)    # Saving our data

#### Checking max lenght of posts

df = data

# Creating a column on the character counts (Code by Brenda Hali)
df['text_characters'] = df['post'].apply(len)
df['text_words'] = df['post'].apply(lambda x: len(x.split()))
df.sort_values('text_words', ascending=False).head(10)
