==> Subreddits NLP/03_Modeling.ipynb <==
%%javascript
IPython.OutputArea.prototype._should_scroll = function(lines) {
return false;
}

import os
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score
from sklearn.naive_bayes import MultinomialNB, BernoulliNB, GaussianNB       # Naive Bayes
from sklearn.linear_model import LogisticRegression
from sklearn import metrics
from sklearn.ensemble import RandomForestClassifier, ExtraTreesClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import VotingClassifier

import warnings
warnings.filterwarnings('ignore')

data = pd.read_csv('./data/data.csv')

df = pd.read_csv('./data/data.csv')

X = data['post']     # Setting the X and y
y = data['subreddit']

y.value_counts(normalize=True)   # Cheking base line 

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, random_state=42)

### Using a pipeline with Multinomial NB and Count Vectorizer

pipeCVnb = Pipeline([           # Using pipeline with CountVectorizer and MultinomialNB
    ('cvect', CountVectorizer()),
    ('Mnb', MultinomialNB())
])

pipeCVnb_params = {         # Setting the parameters 
               'cvect__max_features': [150_000, 200_000], 
               'cvect__stop_words': [None, 'english'],
               'cvect__ngram_range': [(1,3),(1,2)] 
              }

# Using the GridSearchCV to find best parameters
gsCVnb = GridSearchCV(pipeCVnb, # What is the model we want to fit?
                  pipeCVnb_params, # What is the dictionary of hyperparameters?
                  cv=3, verbose=1)

gsCVnb.fit(X_train, y_train)   # Fitting the train data

print(f'Best Score for cv: {gsCVnb.best_score_}')   # Printing best score for CV

print(f'Best Parameters: {gsCVnb.best_params_}') # Checking best parameters for CV

gsCVnb_model = gsCVnb.best_estimator_     # Setting our best model

print(f'Train score: {gsCVnb_model.score(X_train, y_train)}')    # Scoring train data on our best model 

print(f'Test score: {gsCVnb_model.score(X_test, y_test)}')    # Scoring test data on our best model 

os.system("printf '\a'");

# Get report on grid search results
best_model_params = pd.DataFrame(gsCVnb.cv_results_)

best_model_params.to_csv('best_model.csv', index=False)

best_model_params.drop(columns=['mean_fit_time','std_fit_time','mean_score_time','std_score_time',
                               'params', 'split0_test_score','split1_test_score', 'split2_test_score',
                               'std_test_score']).sort_values('rank_test_score')

### Using a pipeline with Multinomial NB and TfidfVectorizer

pipeTIFInb = Pipeline([       # Using pipeline with Tifidifi and MultinomialNB
    ('tfi', TfidfVectorizer(lowercase=False)),
    ('Mnb', MultinomialNB())
])    

pipeTIFInb_params = {       # Setting the parameters
               'tfi__max_features': [75_000],
               'tfi__stop_words': ['english'],
               'tfi__ngram_range': [(1,3),(1,2)]
              }

gsTIFInb = GridSearchCV(pipeTIFInb, # What is the model we want to fit?
                  pipeTIFInb_params, # What is the dictionary of hyperparameters?
                  cv=3, n_jobs=-1, verbose=2)

gsTIFInb.fit(X_train, y_train)   # Fitting the train data

print(f'Best Score for cv: {gsTIFInb.best_score_}')   # Printing best score for CV

print(f'Best Parameters: {gsTIFInb.best_params_}') # Checking best parameters for CV

gsTIFInb_model = gsTIFInb.best_estimator_     # Setting our best model

print(f'Train score: {gsTIFInb_model.score(X_train, y_train)}')    # Scoring train data on our best model 

print(f'Test score: {gsTIFInb_model.score(X_test, y_test)}')    # Scoring test data on our best model 

os.system("printf '\a'");


# Get report on grid search results
pd.DataFrame(gsTIFInb.cv_results_)


### Using a pipeline with Random Forest and Count Vectorizer

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, random_state=42)

pipeCVrf = Pipeline([           # Using pipeline with CountVectorizer and MultinomialNB
    ('cvect', CountVectorizer()),
    ('rf', RandomForestClassifier())
])

pipeCVrf_params = {         # Setting the parameters 
               'cvect__max_features': [50_000], 
               'cvect__stop_words': ['english'],
               'cvect__ngram_range': [(1,1)], 
              }

# Using the GridSearchCV to find best parameters
gsCVrf = GridSearchCV(pipeCVrf, # What is the model we want to fit?
                  pipeCVrf_params, # What is the dictionary of hyperparameters?
                  cv=3, n_jobs=-1)

gsCVrf.fit(X_train, y_train)   # Fitting the train data

print(f'Best Score for cv: {gsCVrf.best_score_}')   # Printing best score for CV

print(f'Best Parameters: {gsCVrf.best_params_}') # Checking best parameters for CV

gsCVrf_model = gsCVrf.best_estimator_     # Setting our best model

print(f'Train score: {gsCVrf_model.score(X_train, y_train)}')    # Scoring train data on our best model 

print(f'Test score: {gsCVrf_model.score(X_test, y_test)}')    # Scoring test data on our best model 



os.system('say -v Samantha Your code is done running');

### Using a pipeline with Random Forest and Tfid Vectorizer

pipeTIFIrf = Pipeline([           # Using pipeline with CountVectorizer and MultinomialNB
    ('tfi', TfidfVectorizer(lowercase=False)),
    ('rf', RandomForestClassifier())
])

pipeTIFIrf_params = {         # Setting the parameters 
               'tfi__max_features': [50_000], 
               'tfi__stop_words': ['english'],
               'tfi__ngram_range': [(1,1)], 
              }

# Using the GridSearchCV to find best parameters
gsTIFIrf = GridSearchCV(pipeTIFIrf, # What is the model we want to fit?
                  pipeTIFIrf_params, # What is the dictionary of hyperparameters?
                  cv=3, n_jobs=-1)

gsTIFIrf.fit(X_train, y_train)   # Fitting the train data

print(f'Best Score for cv: {gsTIFInb.best_score_}')   # Printing best score for CV

print(f'Best Parameters: {gsTIFInb.best_params_}') # Checking best parameters for CV

gsTIFInb_model = gsTIFInb.best_estimator_     # Setting our best model

print(f'Train score: {gsTIFInb_model.score(X_train, y_train)}')    # Scoring train data on our best model 

print(f'Test score: {gsTIFInb_model.score(X_test, y_test)}')    # Scoring test data on our best model 



### Using a pipeline with Extra Trees and Count Vectorizer

pipeCVet = Pipeline([           # Using pipeline with CountVectorizer and MultinomialNB
    ('cvect', CountVectorizer()),
    ('et', ExtraTreesClassifier())
])

pipeCVet_params = {         # Setting the parameters 
               'cvect__max_features': [50_000], 
               'cvect__stop_words': ['english'],
               'cvect__ngram_range': [(1,1)], 
              }

# Using the GridSearchCV to find best parameters
gsCVet = GridSearchCV(pipeCVet, # What is the model we want to fit?
                  pipeCVet_params, # What is the dictionary of hyperparameters?
                  cv=3, n_jobs=-1)

gsCVet.fit(X_train, y_train)   # Fitting the train data

print(f'Best Score for cv: {gsCVet.best_score_}')   # Printing best score for CV

print(f'Best Parameters: {gsCVet.best_params_}') # Checking best parameters for CV

gsCVet_model = gsCVet.best_estimator_     # Setting our best model

print(f'Train score: {gsCVet_model.score(X_train, y_train)}')    # Scoring train data on our best model 

print(f'Test score: {gsCVet_model.score(X_test, y_test)}')    # Scoring test data on our best model 



### Using a pipeline with Extra Trees and Tfid Vectorizer

pipeTIFIet = Pipeline([           # Using pipeline with CountVectorizer and MultinomialNB
    ('tfi', TfidfVectorizer(lowercase=False)),
    ('et', ExtraTreesClassifier())
])

pipeTIFIet_params = {         # Setting the parameters 
               'tfi__max_features': [50_000], 
               'tfi__stop_words': ['english'],
               'tfi__ngram_range': [(1,1)], 
              }

# Using the GridSearchCV to find best parameters
gsTIFIet = GridSearchCV(pipeTIFIet, # What is the model we want to fit?
                  pipeTIFIet_params, # What is the dictionary of hyperparameters?
                  cv=3, n_jobs=-1)

gsTIFIet.fit(X_train, y_train)   # Fitting the train data

print(f'Best Score for cv: {gsTIFIet.best_score_}')   # Printing best score for CV

print(f'Best Parameters: {gsTIFIet.best_params_}') # Checking best parameters for CV

gsTIFIet_model = gsTIFIet.best_estimator_     # Setting our best model

print(f'Train score: {gsTIFIet_model.score(X_train, y_train)}')    # Scoring train data on our best model 

print(f'Test score: {gsTIFIet_model.score(X_test, y_test)}')    # Scoring test data on our best model 



### Using a pipeline with Logistic Regression and Tfid Vectorizer

pipeTIFIlr = Pipeline([           # Using pipeline with TfidVectorizer and Logistic Regression
    ('tfi', TfidfVectorizer(lowercase=False)),
    ('lr', LogisticRegression())
])

pipeTIFIlr_params = {         # Setting the parameters 
               'tfi__max_features': [50_000, 75_000, 100_000], 
               'tfi__stop_words': [None, 'english'],
               'tfi__ngram_range': [(1,1),(1,2)]
              }

# Using the GridSearchCV to find best parameters
gsTIFIlr = GridSearchCV(pipeTIFIlr, # What is the model we want to fit?
                  pipeTIFIlr_params, # What is the dictionary of hyperparameters?
                  cv=3, n_jobs=-1)

gsTIFIlr.fit(X_train, y_train)   # Fitting the train data

print(f'Best Score for cv: {gsTIFIlr.best_score_}')   # Printing best score for CV

print(f'Best Parameters: {gsTIFIlr.best_params_}') # Checking best parameters for CV

gsTIFIlr_model = gsTIFIlr.best_estimator_     # Setting our best model

print(f'Train score: {gsTIFIlr_model.score(X_train, y_train)}')    # Scoring train data on our best model 

print(f'Test score: {gsTIFIlr_model.score(X_test, y_test)}')    # Scoring test data on our best model 



# Get report on grid search results
best_model_params = pd.DataFrame(gsTIFIlr.cv_results_)

best_model_params

best_model_params.to_csv('best_model.csv', index=False)

#### Trying out my best model with 115 rows of new data

math_for_predictions = pd.read_csv('math_for_predictions.csv')

A = math_for_predictions['post']

A.shape

gsTIFIlr_model.predict(A)

A[0]

### Using a pipeline with Logistic Regression and Count Vectorizer


pipeCVlr = Pipeline([           # Using pipeline with CountVectorizer and Logistic Regression
    ('cvect', CountVectorizer()),
    ('lr', LogisticRegression())
])

pipeCVlr_params = {         # Setting the parameters 
               'cvect__max_features': [100_000, 200_000], 
               'cvect__stop_words': ['english'],
               'cvect__ngram_range': [(1,2),(1,3)], 
              }

# Using the GridSearchCV to find best parameters
gsCVlr = GridSearchCV(pipeCVlr, # What is the model we want to fit?
                  pipeCVlr_params, # What is the dictionary of hyperparameters?
                  cv=3, verbose=1)

gsCVlr.fit(X_train, y_train)   # Fitting the train data

print(f'Best Score for cv: {gsCVlr.best_score_}')   # Printing best score for CV

print(f'Best Parameters: {gsCVlr.best_params_}') # Checking best parameters for CV

gsCVlr_model = gsCVlr.best_estimator_     # Setting our best model

print(f'Train score: {gsCVlr_model.score(X_train, y_train)}')    # Scoring train data on our best model 

print(f'Test score: {gsCVlr_model.score(X_test, y_test)}')    # Scoring test data on our best model 



### Using a pipeline with Vote Classifier and Count Vectorizer


pipeTIFIet = Pipeline([           # Using pipeline with Vote Classifier
    ('tfi', TfidfVectorizer(lowercase=False)),
    ('vote', VotingClassifier([
    ('lr', LogisticRegression()),
    ('MNB', MultinomialNB()),
    ('et', ExtraTreesClassifier()),
    ('rf', RandomForestClassifier())]))
])

pipeTIFIet_params = {         # Setting the parameters 
               'tfi__max_features': [75_000, 100_000], 
               'tfi__stop_words': ['english'],
               'tfi__ngram_range': [(1,1),(1,2)], 
              }

# Using the GridSearchCV to find best parameters
gsTIFIet = GridSearchCV(pipeTIFIet, # What is the model we want to fit?
                  pipeTIFIet_params, # What is the dictionary of hyperparameters?
                  cv=3, verbose=1)

gsTIFIet.fit(X_train, y_train)   # Fitting the train data

print(f'Best Score for cv: {gsTIFIet.best_score_}')   # Printing best score for CV

print(f'Best Parameters: {gsTIFIet.best_params_}') # Checking best parameters for CV

gsTIFIet_model = gsTIFIet.best_estimator_     # Setting our best model

print(f'Train score: {gsTIFIet_model.score(X_train, y_train)}')    # Scoring train data on our best model 

print(f'Test score: {gsTIFIet_model.score(X_test, y_test)}')    # Scoring test data on our best model
