==> Subreddits NLP/04_Code_for_Visualizations.ipynb <==
### Creating a DataFrame to store our results and visualize them

Scores = pd.DataFrame(columns=['Model','Train Score', 'Test Score'])

Scores['Model'] = ['Multinomial Naive Bayes with Count Vectorizer',
                   'Multinomial Naive Bayes with Tfid Vectorizer', 
                   'Random Forest with Count Vectorizer', 
                   'Random Forest with Tfid Vectorizer', 
                   'Extra Trees with Count Vectorizer', 
                   'Extra Trees with Tfid Vectorizer',
                   'Logistic Regression with Count Vectorizer',
                   'Logistic Regression with Tfid Vectorizer']

Scores['Train Score'] = [0.9415257020857284, 
                         0.9311299063885695, 
                         0.9944818525209393, 
                         0.9280916406634916, 
                         0.998127771391033, 
                         0.9984233864345541,
                         0.984455575628182,
                         0.9551650517326327] 


Scores['Test Score'] = [0.9259496477311918, 
                        0.9179435384539587, 
                        0.9043701039562497, 
                        0.9170567078878652, 
                        0.8979405823520717, 
                        0.9058481548997389,
                        0.9389564960338966,
                        0.9389072276691136] 

Scores

#### Saving it to use in tableau

Scores.to_csv('scores.csv', index=False)

### Importing our best parameters from Modeling for visualization

best_model_params = pd.read_csv('best_model.csv')

best_model_params

#### Dropping everything we do not really need to visualize and ordering by rank

best_model_params.drop(columns=['mean_fit_time','std_fit_time','mean_score_time','std_score_time',
                               'params', 'split0_test_score','split1_test_score', 'split2_test_score',
                               'std_test_score']).sort_values('rank_test_score')
